+++ src/app/src/main/java/jp/kaleidot725/emomemo/ui/memo/AddMemoDialogFragment.kt
@@ -1,4 +1,4 @@
-package jp.kaleidot725.emomemo.ui.homedialog
+package jp.kaleidot725.emomemo.ui.memo
 
 import android.os.Bundle
 import android.view.View
@@ -7,17 +7,17 @@ import androidx.lifecycle.Observer
 import androidx.navigation.NavController
 import androidx.navigation.fragment.findNavController
 import jp.kaleidot725.emomemo.R
-import jp.kaleidot725.emomemo.databinding.FragmentHomeDialogBinding
+import jp.kaleidot725.emomemo.databinding.FragmentAddMemoBinding
 import jp.kaleidot725.emomemo.extension.viewBinding
 import jp.kaleidot725.emomemo.ui.MainViewModel
 import org.koin.androidx.viewmodel.ext.android.sharedViewModel
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class HomeDialogFragment : DialogFragment(R.layout.fragment_home_dialog) {
+class AddMemoDialogFragment : DialogFragment(R.layout.fragment_add_memo) {
     private val navController: NavController get() = findNavController()
-    private val binding: FragmentHomeDialogBinding by viewBinding()
+    private val binding: FragmentAddMemoBinding by viewBinding()
     private val mainViewModel: MainViewModel by sharedViewModel()
-    private val homeDialogViewModel: HomeDialogViewModel by viewModel()
+    private val addMemoViewModel: AddMemoViewModel by viewModel()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -27,14 +27,14 @@ class HomeDialogFragment : DialogFragment(R.layout.fragment_home_dialog) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         binding.mainViewModel = mainViewModel
-        binding.homeDialogViewModel = homeDialogViewModel
-        homeDialogViewModel.event.observe(viewLifecycleOwner, Observer { event ->
+        binding.addMemoViewModel = addMemoViewModel
+        addMemoViewModel.event.observe(viewLifecycleOwner, Observer { event ->
             when (event) {
-                HomeDialogViewModel.NavEvent.SUCCESS -> {
-                    mainViewModel.createMemo(homeDialogViewModel.title.value ?: "")
+                AddMemoViewModel.NavEvent.SUCCESS -> {
+                    mainViewModel.createMemo(addMemoViewModel.title.value ?: "")
                     navController.popBackStack()
                 }
-                HomeDialogViewModel.NavEvent.CANCEL -> {
+                AddMemoViewModel.NavEvent.CANCEL -> {
                     navController.popBackStack()
                 }
                 else -> {
