+++ src/app/src/main/java/jp/kaleidot725/emomemo/ui/memo/AddMemoViewModel.kt
@@ -1,16 +1,11 @@
-package jp.kaleidot725.emomemo.ui.homedialog
+package jp.kaleidot725.emomemo.ui.memo
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.hadilq.liveevent.LiveEvent
-import jp.kaleidot725.emomemo.model.AppStatus
-import jp.kaleidot725.emomemo.model.db.repository.MemoRepository
 
-class HomeDialogViewModel(
-    private val appStatus: AppStatus,
-    private val memoRepository: MemoRepository
-) : ViewModel() {
+class AddMemoViewModel : ViewModel() {
     private val _event: LiveEvent<NavEvent> = LiveEvent()
     val event: LiveData<NavEvent> = _event
     val title: MutableLiveData<String> = MutableLiveData()
