+++ src/app/src/main/res/layout/fragment_add_memo.xml
@@ -8,8 +8,8 @@
             type="jp.kaleidot725.emomemo.ui.MainViewModel" />
 
         <variable
-            name="homeDialogViewModel"
-            type="jp.kaleidot725.emomemo.ui.homedialog.HomeDialogViewModel" />
+            name="addMemoViewModel"
+            type="jp.kaleidot725.emomemo.ui.memo.AddMemoViewModel" />
     </data>
 
     <com.google.android.material.card.MaterialCardView xmlns:android="http://schemas.android.com/apk/res/android"
@@ -42,7 +42,7 @@
                 android:gravity="start|top"
                 android:hint="@string/home_dialog_fragment_memo_title_hint"
                 android:lines="2"
-                android:text="@={homeDialogViewModel.title}"
+                android:text="@={addMemoViewModel.title}"
                 android:textColor="@android:color/black"
                 app:layout_constraintTop_toBottomOf="@id/title_text_view" />
 
@@ -52,7 +52,7 @@
                 android:layout_height="wrap_content"
                 android:layout_margin="8dp"
                 android:backgroundTint="@color/primaryColor"
-                android:onClick="@{() -> homeDialogViewModel.success()}"
+                android:onClick="@{() -> addMemoViewModel.success()}"
                 android:text="@string/home_dialog_fragment_create"
                 android:textColor="@color/primaryTextColor"
                 app:layout_constraintEnd_toStartOf="@id/cancel_button"
@@ -65,7 +65,7 @@
                 android:layout_height="wrap_content"
                 android:layout_margin="8dp"
                 android:backgroundTint="@color/primaryColor"
-                android:onClick="@{() -> homeDialogViewModel.cancel()}"
+                android:onClick="@{() -> addMemoViewModel.cancel()}"
                 android:text="@string/home_dialog_fragment_cancel"
                 android:textColor="@color/primaryTextColor"
                 app:layout_constraintEnd_toEndOf="parent"
